Option `--tagq` qualifies the build with the Tag used to fetch the src.
We changed how we version the release, you now have to explicitly use
the `--tagq` option to qualify the build with the tag passed through the
`--tag` option. If no qualifier is specified we will default to the git
hash.

Examples

If you specify the build qualifier as *rc1* and the tag as *0.18.0-rc1*
you will get the rpms with version *mesos-0.18.0-rc1*.

```
./mesos-build --qualifier=rc1 --tag=0.18.0-rc1
```

If you don't specify the build qualifier but you do specify a tag, let say *0.18.0-rc1*
you will get the rpms with version *mesos-0.18.0-h_57b57a2*. Where *h_57b57a2* is the
git hash that belongs to the given tag.

```
./mesos-build --tag=0.18.0-rc1
```

[ticket: X] : ?